backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
aiohttp==3.9.1
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]
bcrypt==4.0.1
ratelimit==2.2.1